convert embed to fn, move convertToDelimited
